refactor: files in domain in specfic foldes, fix imports
